Recycle explorer menu tests
